Add permissions for persistent volume claims
We use `aws:RequestTag` to ensure that any calls to `ec2:CreateTags`
from the controller can only set the `KubernetesCluster` tag to the
cluster name. This means that we can reasonably sure that a controller
in one environment can't change tags in order to attach a volume from a
different environment.

Based off the permissions from here[1]. We've only added things that
we've seen be required. We have only added permissions that we've seen
failed Unauthorized calls for in CloudTrail:

```
$ aws cloudtrail lookup-events \
  --start-time 2019-03-18T12:00:00Z \
  | jq -r '.Events[].CloudTrailEvent | fromjson | select(.userIdentity.arn | test("samcrang"))'
```

[1] https://docs.docker.com/ee/ucp/kubernetes/storage/configure-aws-storage/